+++ src/components/Details/TranscriptSearch/TranscriptSearch.stories.tsx
@@ -8,6 +8,8 @@ import { initialFetchDataState } from "../../../containers/FetchDataContainer/us
 
 import TranscriptSearch, { TranscriptSearchProps } from "./TranscriptSearch";
 
+import { mockSentences } from "../../../stories/model-mocks/sentence";
+
 export default {
   component: TranscriptSearch,
   title: "Library/Details/Transcript Search",
@@ -27,27 +29,7 @@ Default.args = {
   searchQuery: "sentence",
   sentences: {
     ...initialFetchDataState,
-    data: [
-      {
-        session_index: 0,
-        index: 0,
-        start_time: 1.5,
-        text: "This is a sentence.",
-        speaker_index: 0,
-        speaker_name: "Lisa Herbold",
-        speaker_id: "lisa-herbold",
-        speaker_pictureSrc:
-          "https://www.seattle.gov/images/Council/Members/Herbold/Herbold_225x225.jpg",
-      },
-      {
-        session_index: 0,
-        index: 1,
-        start_time: 2,
-        text: "This is another sentence.",
-        speaker_name: "Unknown",
-        speaker_index: 1,
-      },
-    ],
+    data: mockSentences(1, 2),
   },
   jumpToVideoClip: (sessionIndex, startTime) => action("jump to video clip"),
   jumpToTranscript: (sentenceIndex) => action("jump to transcript"),
@@ -58,17 +40,7 @@ manySentences.args = {
   searchQuery: "sentence",
   sentences: {
     ...initialFetchDataState,
-    data: Array.from({ length: 200 }).map((_, i) => ({
-      session_index: 0,
-      index: i,
-      start_time: i,
-      text: `This is a sentence ${i}.`,
-      speaker_name: "Lisa Herbold",
-      speaker_id: "lisa-herbold",
-      speaker_index: 0,
-      speaker_pictureSrc:
-        "https://www.seattle.gov/images/Council/Members/Herbold/Herbold_225x225.jpg",
-    })),
+    data: mockSentences(2, 100),
   },
   jumpToVideoClip: (sessionIndex, startTime) => action("jump to video clip"),
   jumpToTranscript: (sentenceIndex) => action("jump to transcript"),
